+++ src/app/models/big-mall.model.ts
@@ -1,16 +1,4 @@
-import { BasicBusinessModel } from './common/basicBusinessModel.model';
+import { BasicBusinessModel } from './common/basic-business.model';
 
 export interface BigMallModel extends BasicBusinessModel {
-    addresses: Address[]
-}
-interface Address {
-    address_name: string;
-    street_number: string;
-    zip_code: string;
-    district_id: string;
-    town: string;
-    location: {
-        x: number;
-        y: number
-    }
 }
