--- src/app/models/common/basicBusinessModel.model.ts
@@ -1,9 +1,0 @@
-import { EconomicActivityModel } from "./economic-activity.model";
-
-export interface BasicBusinessModel {
-    name: string;
-    web: string | null;
-    email: string | null;
-    phone: number | null;
-    activities: EconomicActivityModel[];
-}
