+++ src/assets/dummy/big-malls_dummy.json
@@ -245,8 +245,7 @@
                 },
                 {
                     "id": 32798724,
-                    "name": "Discos",
-                    "full_path": "Productes >> Imatge i so >> Discos"
+                    "name": "Discos"
                 },
                 {
                     "id": 107007,
@@ -550,8 +549,7 @@
                 },
                 {
                     "id": 66013073,
-                    "name": "Otras maquinas automaticas",
-                    "full_path": "Arbre Principal Barcelona Activa  >> Industria, comerc i serveis >> Otros servicios >> Servicios recreativos y culturales >> Otros servicios recreativos n.c.o.p. >> Otras maquinas automaticas"
+                    "name": "Otras maquinas automaticas"
                 }
             ],
             "addresses":
